Rename files to conform to standard Python practice
